feat: restart app automatically when update is complete
